Move feeCalc to execution folder
